+++ crates/radiate-python/src/bindings/genome/species.rs
@@ -119,7 +119,7 @@ macro_rules! impl_into_py_species {
 }
 
 impl_into_py_species!(FloatChromosome);
-impl_into_py_species!(IntChromosome<i32>);
+impl_into_py_species!(IntChromosome<i64>);
 impl_into_py_species!(BitChromosome);
 impl_into_py_species!(CharChromosome);
 impl_into_py_species!(GraphChromosome<Op<f32>>);
